Move gogocodec to internal
