+++ internal/gogocodec/codec.go
